requirements.txt
asgiref==3.2.5
Babel==2.8.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-phonenumber-field==4.0.0
idna==2.9
oauthlib==3.1.0
phonenumberslite==8.11.5
Pillow==7.0.0
pkg-resources==0.0.0
pycparser==2.20
pycryptodome==3.9.7
pydotenv==0.0.7
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
